tablets: join co-located tablet during migration
A co-located table may be allocated while a base tablet is migrating.
In that case, the new tablet is put in the "join_group" transition stage
and it has no replicas.

When the migration is completed, the joining tablets exit the join_group
stage and their replicas are set, and storage is allocated.