sp: handle replica_local in mutate
This commit introduces the replica_local flag, which forces a query
to be executed only on the local replica. This flag is needed for
Paxos state access, where only local Paxos state must be consulted.
In the future, we may introduce a CQL extension to expose this
"local mode" to users, which could be useful for debugging
and troubleshooting.

We also add the remove_non_local_host_ids() helper, which will be used
in upcoming commits to support the read path.

The storage_proxy_coordinator_mutate_options class is declared outside
of storage_proxy to avoid C++ compiler complaints about default field
initializers. In particular, some storage_proxy methods use this class
for optional parameters with default values, which is not allowed when
the class is defined inside storage_proxy.